绘本，文章详情bug修复

+++ app/src/main/java/com/dace/textreader/bean/AudioArticleBean.java
@@ -132,7 +132,7 @@ public class AudioArticleBean {
              * mean : null
              */
 
-            private int sequence;
+            private long sequence;
             private int id;
             private String type;
             private String title;
@@ -179,11 +179,11 @@ public class AudioArticleBean {
             private List<ContentListBean> contentList;
             private List<?> annotation;
 
-            public int getSequence() {
+            public long getSequence() {
                 return sequence;
             }
 
-            public void setSequence(int sequence) {
+            public void setSequence(long sequence) {
                 this.sequence = sequence;
             }
